Integrate profiles' switcher (#374).
Also, fix typos in raws location & don't make the core depend on the IrcConnector module
when disconnecting (to retrieve the quit message). Instead, the bot now dispatches
an Erebot_Event_Exit event (as was already the case), which is captured by that module
to force a disconnection.
This makes the core more resilient to (future) changes in Erebot_Module_IrcConnector's API.